Extract methods to get the table names to extract when doing a snapshot
